Fixed new getPhotoById route
