+++ backend/src/main/kotlin/gallery/kurate/server/api/handler/PhotoHandlers.kt
@@ -2,13 +2,16 @@ package gallery.kurate.server.api.handler
 
 import co.selim.gimbap.api.StreamingStore
 import gallery.kurate.server.api.NotFoundException
+import gallery.kurate.server.api.buildImageUrl
 import gallery.kurate.server.database.AlbumRepository
 import gallery.kurate.server.database.PhotoRepository
 import gallery.kurate.server.ext.dispose
 import gallery.kurate.server.ext.endWithJson
 import gallery.kurate.server.ext.notFound
 import io.reactivex.Maybe
+import io.reactivex.Observable
 import io.reactivex.Single
+import io.reactivex.internal.operators.observable.ObservableError
 import io.reactivex.schedulers.Schedulers
 import io.vertx.core.Handler
 import io.vertx.core.json.JsonObject
@@ -16,6 +19,7 @@ import io.vertx.kotlin.core.json.jsonArrayOf
 import io.vertx.reactivex.core.buffer.Buffer
 import io.vertx.reactivex.ext.web.RoutingContext
 import java.util.*
+import java.util.logging.Logger
 
 data class ImageHolder(val imageBytes: ByteArray? = null)
 
@@ -42,19 +46,19 @@ internal fun getImageHandler(imageStore: StreamingStore<ByteArray>) =
   }
 
 internal fun getPhotoByIdHandler(
-  photoRepository: PhotoRepository
+  photoRepository: PhotoRepository,
+  host: String
 ) =
   Handler<RoutingContext> { context ->
     val photoId = context.request().getParam("id")
-    val disposable = photoRepository.getPhotoById(photoId)
+    val disposable = photoRepository.getPhotoById(photoId).toObservable()
       .subscribeOn(Schedulers.io())
-      .switchIfEmpty(Maybe.error(NotFoundException("Photo with id $photoId not foundx")))
+      .switchIfEmpty(Observable.error(NotFoundException("Photo with id $photoId not found")))
       .doOnError(context::fail)
-      .map { photo ->
-        photo.getJsonArray("photos", jsonArrayOf())
-        photo
+      .map { res ->
+        res.getJsonObject("photos").also { populateUrls(it, host) }
       }
-      .subscribe { album -> context.endWithJson(album) }
+      .subscribe { photo -> context.endWithJson(photo) }
 
     context.dispose(disposable)
   }
