+++ backend/src/main/kotlin/gallery/kurate/server/database/PhotoRepository.kt
@@ -1,14 +1,17 @@
 package gallery.kurate.server.database
 
 import io.reactivex.Maybe
+import io.vertx.core.json.JsonArray
 import io.vertx.core.json.JsonObject
+import io.vertx.kotlin.core.json.jsonArrayOf
 import io.vertx.kotlin.core.json.jsonObjectOf
+import io.vertx.reactivex.core.streams.ReadStream
 import io.vertx.reactivex.ext.mongo.MongoClient
 import org.bson.types.ObjectId
 
 interface PhotoRepository {
   fun addPhoto(albumId: String, fileName: String, uri: String): Maybe<JsonObject>
-  fun getPhotoById(id: String): Maybe<JsonObject>
+  fun getPhotoById(id: String): ReadStream<JsonObject>
 }
 
 private const val collection = "albums"
@@ -20,8 +23,15 @@ class MongoPhotoRepository(private val mongoClient: MongoClient) : PhotoReposito
       .map { jsonObjectOf("_id" to photoId) }
   }
 
-  override fun getPhotoById(id: String): Maybe<JsonObject> = mongoClient
-    .rxFindOne(collection, jsonObjectOf("photos._id" to id), null)
+  override fun getPhotoById(id: String): ReadStream<JsonObject> = mongoClient
+    .aggregate(
+      collection, jsonArrayOf(
+        jsonObjectOf("\$match" to jsonObjectOf("photos._id" to id)),
+        jsonObjectOf("\$unwind" to jsonObjectOf("path" to "\$photos")),
+        jsonObjectOf("\$match" to jsonObjectOf("photos._id" to id)),
+        jsonObjectOf("\$project" to jsonObjectOf("photos" to 1, "_id" to 0))
+      )
+    );
 
   private fun doAddPhoto(
     albumId: String,
